unmount averything thats mounted into the chroot
Installing certain packages may actually mount things inside the chroot,
e.g. /proc/sys/fs/binfmt_misc. This needs to be taken care of when
removing the chroot.

Add to Chroot two methods get_mounted() and unmount_all() that replace
unmount_proc(). Use them for cleanup and add some extra checking before
running rmtree().

TODO: Cooperation + order with unmount_selinux() is unclear.
TODO: LVM and VirtServ support?

TODO: Possible problems: symlinks on the way to the tempdir may cause
/proc/mounts to use a different directory name from Chroot.name, so that
prefix matching does not work and the mount points are not found.
Assuming "everything unmounted" in that case may be problematic.

A rather safe approach to cleanup the chroot would be
* umount known extra mounts (e.g. /proc/sys/fs/binfmt_misc)
* umount all known mountpoints (/proc, bindmounts)
* umount everything recorded in /proc/mounts matching this chroot
* use rm -rf --one-file-system

+++ piuparts.py
@@ -785,13 +785,16 @@ class Chroot:
         if not settings.keep_tmpdir and os.path.exists(self.name):
             self.terminate_running_processes()
             self.unmount_selinux()
-            self.unmount_proc()
+            self.unmount_all()
             if settings.lvm_volume:
                 logging.debug('Unmounting and removing LVM snapshot %s' % self.lvm_snapshot_name)
                 run(['umount', self.name])
                 run(['lvremove', '-f', self.lvm_snapshot])
-            shutil.rmtree(self.name)
-            logging.debug("Removed directory tree at %s" % self.name)
+            if self.get_mounted():
+                logging.error("Can't remove directory tree at %s, the following directories are still mounted: %s" % (self.name, " ".join(self.get_mounted())))
+            else:
+                shutil.rmtree(self.name)
+                logging.debug("Removed directory tree at %s" % self.name)
         elif settings.keep_tmpdir:
             logging.debug("Keeping directory tree at %s" % self.name)   
 
@@ -1268,6 +1271,24 @@ class Chroot:
         self.run(["umount", "/proc"], ignore_errors=True)
         for bindmount in settings.bindmounts:
             run(["umount", self.relative(bindmount)], ignore_errors=True)
+
+    def get_mounted(self):
+        mounts = []
+        proc_mounts = open("/proc/mounts", "r")
+        try:
+            for line in proc_mounts:
+                mountpoint = line.split()[1]
+                if mountpoint[:len(self.name) + 1] == self.name + "/":
+                    mounts.append(mountpoint[len(self.name):])
+        finally:
+            proc_mounts.close()
+        mounts.reverse()
+        return mounts
+
+    def unmount_all(self):
+        """Unmount everything inside chroot."""
+        for mountpoint in self.get_mounted():
+            self.run(["umount", mountpoint], ignore_errors=True)
 
     def is_ignored(self, pathname):
         """Is a file (or dir or whatever) to be ignored?"""
